refactor: Use ts for next config
